[PLIST] Create data source and interactor for getting posts, configure DI.
- Configure DI manually by choice
- Create mappers
- Update models types from Int to Long
- Add ContentEntity to PostEntity

+++ app/src/main/java/com/eagskunst/topddit/data/model/ChildrenDataModel.kt
@@ -34,9 +34,9 @@ data class ChildrenDataModel(
     @SerialName("contest_mode")
     val contestMode: Boolean,
     @SerialName("created")
-    val created: Int,
+    val created: Long,
     @SerialName("created_utc")
-    val createdUtc: Int,
+    val createdUtc: Long,
     @SerialName("domain")
     val domain: String,
     @SerialName("downs")
